server/requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==44.0.2
Cython==3.0.12
Django==5.1.1
django_environ==0.12.0
djangorestframework==3.16.0
dl==0.1.0
docutils==0.21.2
drf_yasg==1.21.10
environ==1.0
filelock==3.18.0
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.6
keyring==25.6.0
mock==5.2.0
Pillow==11.2.1
protobuf==6.30.2
pyOpenSSL==25.0.0
railroad==0.5.0
redis==5.2.1
Sphinx==8.2.3
thread==2.0.5
tornado==6.4.1
djangorestframework-simplejwt=5.5.0
django-debug-toolbar==5.1.0
psycopg2==2.9.0
django-cors-headers==4.7.0
